pyproject.toml
[project]
name = "brook-agent"
version = "0.1.0"
description = "Minimal LLM agent MVP using agentloop + OpenAI-compatible API"
requires-python = ">=3.10"
dependencies = [
  "agentloop>=0.2.0,<0.3",
  "openai>=1.0.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
brook-agent = "brook_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["brook_agent"]
